% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_df_crossref}
\alias{extract_df_crossref}
\title{extract_df_crossref}
\usage{
extract_df_crossref(dois)
}
\arguments{
\item{dois}{a list of dois}
}
\value{
a data fraem
}
\description{
extract_df_crossref
}
\examples{
dois <- c("10.5284/1011335","10.1126/science.169.3946.635")
extract_df_crossref(dois)

}
