% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_dois}
\alias{find_dois}
\title{find_dois}
\usage{
find_dois(query, limit = 1)
}
\arguments{
\item{query}{query}

\item{limit}{number papers to extract}
}
\value{
a data fraem
}
\description{
find_dois
}
\examples{
query <- "Survival in frontotemporal lobar degeneration and related disorders: latent class predictors and brain functional correlates"
find_dois(query = query)

queryList <- c("COVID-19 and Myocarditis: What Do We Know So Far?", "Neurologic complications of COVID-19")
sapply(queryList, find_dois)

}
