% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_dois}
\alias{find_dois}
\title{find_dois}
\usage{
find_dois(
  input_title,
  input_year = NA,
  limit = 2,
  title_distance_threshold = 0.93
)
}
\arguments{
\item{input_title}{input title to serach for}

\item{input_year}{input year to search for}

\item{limit}{number papers to extract}

\item{title_distance_threshold}{a double number. The treshold of the distance between true title and false}
}
\value{
a data fraem
}
\description{
find_dois
}
\examples{
input_title <- "Survival in frontotemporal lobar degeneration and related disorders"
input_year <- 2009
find_dois(input_title, input_year)

}
